+++ systemimage/bag.py
@@ -84,6 +84,23 @@ class Bag:
     def __getitem__(self, key):
         return self.__untranslated__[key]
 
+    def keys(self):
+        # Like `dict.keys()` but only here instead of the base class because
+        # I'm not sure this is generally useful.
+        for key in self.__dict__:
+            if not key.startswith('_'):
+                yield key
+
+    def get(self, key, default=None):
+        if key in self.__dict__:
+            return self.__dict__[key]
+        return default
+
+    def __iter__(self):
+        # Like `iter(dict)` but only here instead of the base class because
+        # I'm not sure this is generally useful.
+        yield from self.keys()
+
     # Pickle protocol.
 
     def __getstate__(self):